Refactor `vector` module into submodules
It was previously quite large and contained several types/traits, so
refactoring makes it easier to find them.
